fix: Resolve test compilation errors across multiple modules
- Fix SearchResult and SearchVectorRecord imports in cache tests
- Add missing fields to FilterableColumnStats initializers
- Implement missing methods in AccessPatternTracker
- Fix ColumnarSearchConfig field issues in NOVA engine
- Add missing SearchParams fields in Parquet reader tests
- Fix SqlValue enum variant usage
- Correct method names and type conversions
- Update Entity struct field names and types
- Fix P2Matrix type mismatches in Raptor tests
- Address string comparison and cache metrics issues

+++ src/storage/engines/impls/raptor/tests.rs
@@ -471,7 +471,7 @@ mod tests {
             let mut transposed_data = vec![0.0f32; num_centroids * dimension];
 
             for i in 0..num_centroids {
-                rowgroup_ids.push(i as u32);
+                rowgroup_ids.push(i as u16);
 
                 // Create test pattern for centroids
                 for dim in 0..dimension {
@@ -482,7 +482,7 @@ mod tests {
 
             let columnar = ColumnarCentroids {
                 count: num_centroids as u32,
-                dimension: dimension as u32,
+                dimension: dimension,
                 rowgroup_ids,
                 transposed_data,
                 encoding_metadata: vec![],
